API: Change `summary` endpoint from GET to POST
This API should mirror the `route` API, which is also a POST. While they
could be GET API, as the same parameters expect to return the same data
and it does not trigger any data update in the system, the fact that
they pass geojson features makes it easier to use a POST request. For a
GET, ideally, only the coordinates should be returned.