transitRouting: Move TransitRoutingService to the backend
fixes #770

The `TransitRoutingService` class is moved to the backend and instead of
using the socket/fetch routes, it wraps the calls to the trRouting
backend to return the results.

The current implementation using socket or fetch route is not required
anymore, the direct socket routes for trRouting's route and
accessibility map functions are not required anymore and have been
removed.

Remove the `TransitRoutingServiceManager` class, which is not
necessary anymore.