transit: Rename chaire-lib-common's trRouting service to transitRouting
trRouting refers to the `trRouting` project, but the service itself,
while currently heavily coupled with trRouting provides transit routing
in general and trRouting should only be an implementation detail.
Renaming to transitRouting makes the distinction clearer.

This commit only renames the directories and packages, as later commits
will refactor the classes themselves and rename them properly at that
time.