accessibilityMap: Bring the calculation to the backend
The main accessibility map calculation function is now moved to the
backend instead of the common package.

This adds a socket route `accessibilityMap.calculateWithPolygons` to let
the frontend have access to the calculation with polygon.

This will lead to later PRs removing the TrRoutingServiceManager, which
is only used in the accessibility map calculations.