debug: Remove log statements from `TrRoutingServiceBackend`'s route
reverts changes done in this file in 657976c2b9e1d3a828514125a28d4c89e4fc989f

This log statements were wrongly placed as the request is a promise and
the log happened before awaiting for the promise completion.

Also, with the TransitRoutingService in the backend, this log is not
necessary anymore, as the caller already has log statements.