TrRoutingConfig: Add log files and debug configuration options
This adds a `debug` and `logs` configuration option to the
TrRoutingConfig type. They can be configured on the `config.js` used in
the project.

`debug` is a boolean option specifying whether to start the trRouting
process in debug mode

`logs` configures the loggins of the trRouting process. Currently, the
options that can be set are `maxFileSizeKB` for the maximum size of a
log file in kilobytes and `nbFiles` to specify the number of log files
to keep.

They can be set separately for the batch or single instance. These
onfigurations are optional.

Move the default log file count and size to the config instead of the
`ProcessManager.ts` file.