schedules in db: add a query to duplicate whole schedules
This query receives a line, service and path ID mapping and copy the
schedules for those objects in 3 sql queries (one for each table:
schedules, periods and trips).

The query to copy in a single statement per table for all schedules is
inspired by these 2 articles from stack overflow:
https://stackoverflow.com/questions/29256888/insert-into-from-select-returning-id-mappings
and
https://dba.stackexchange.com/questions/46410/how-do-i-insert-a-row-which-contains-a-foreign-key
and uses postgres's `with` statement to get the mapping between old and
new IDs of foreign key objects.