trRouting: Move port and cacheAllScenarios to instance specific config
This introduces typing for the chaire-lib-backend's server
configuration. We add a `routing` property to the server config type.
This property will configure each routing or transit mode by specifying
the default engine to use and the available engines' configuration.

It specifies the type for the `trRouting` engine of the `transit` mode. It
contains data for the `single` instance for individual calculations (UI
and public API) and `batch` instance (for tasks). For each instance, the
`port` and `cacheAllScenarios` properties can be set.

The configuration objects replace the `trRouting` preference that
contained the `port` and `batchPortStart` and the current
`trRoutingCacheAllScenarios` configuration option.

Port management when starting or querying trRouting is updated to match
the new configuration options.

Deprecate the current `trRoutingCacheAllScenarios` configuration option,
add a warning when it's being used and if the cacheAllScenarios property
is not set in the config file for the single instance, this value is
taken.