default db: Support ID parameter either string or number
As some tables will migrate to numeric IDs and given that the knex
queries automatically deal with parameter types, we can safely support
either string or number IDs in the default DB functions.

If it is a string, it is expected to be a uuid and we still validate its
type.