schedules: Use numeric ids for schedules tables primary keys
Change the schedule, schedule_periods and schedule_period_trips's
primary id columns from uuid to auto-increment numeric IDs. The old
uuid is kept is a column called `uuid`. The foreign keys are modified
to use the numeric key instead. The `schedule_id` column of the trips'
table is removed, as there is already a schedule_period_id foreign which
has a foreign key to the schedule_id. Having both in trips open the door
to data integrity issues where the period could not belong to the same
schedule as the trip.

auto-increment numeric IDs are more performant than uuids as keys. It
will also allow to more easily duplicate multiple elements in a table
with predictable sort order, that uuids do not allow.

This commit only changes the database table queries. The Schedule class
still use the uuid as `id` as this would require more refactoring of the
base classes, history tracker, forms and client/server messaging, that
will be done in later commits.

The numeric IDs are kept in the `integer_id` fields of the Schedule. The
foreign key are changed to map to the numeric ID.

The impact is "minimal" as the schedules are used mostly in calculations
by trRouting and, in the frontend, are [re-]loaded on demand when
changes occur on a specific line. There is no collection of this object
type where the change of ID could cause issues.