Migrate OSRM settings from preferences to server.config
The settings for the OSRM modes are moved from preferences to config, and use a new format that is more similar to the trRouting modes.

Instead of all the modes that use OSRM being placed in the same `defaultPreferences.osrmRouting.modes` object, they are all placed in the server configuration's `routing` object.
There, all modes no matter the engine are at the same level, have a `defaultEngine` field, and an `engines` field that allows for each mode to have configs for multiple routing engines.
For exemple, at the moment, 'driving' has osrmRouting as its `defaultEngine` and only configured engine, but we could add a new config that uses a different engine without having to create a new `driving` object elsewhere.
We also change the `osrmPath` field to `host`, as it was not used anywhere in the code, and we can now use `host` to use remote OSRM instances.

At the moment, we still allow the old format to be used so that the instances aren't broken, but it will be fully deprecated in the future.
Fix: #1084