doc tooltip: Fetch on demand and cache definition results
The `DocumentationTooltip` component now requests the definition only if
the user clicks to display the tooltip. The loader is shown while it is
loading.

The `DefinitionService` caches the results from previous server calls,
such that it is loaded only once. If a specific definition request is
ongoing, it does not call the server again but waits for the server to
return a value or error. Upon error, the next call to get the definition
will try to get it again.

Add unit tests for the `getDefinitionFromServer` function