path: let the path geojson collection db query return the features
The `jsonb_agg` function that aggregates the features of the path
collection to return a geojson collection is not very performant. On a
large network (the whole Montreal area), it multiplies by 5 the
execution time calcultated in the `explain analyze` query, compared to
the query without the aggregation (just returning the features).

So instead, the database query will return the json features and
typescript will do the coalescing in a geojson collection.

The result of the database query before was a single row with a geojson
feature collection object as follows:

```
row 1: {
  "type": "FeatureCollection",
  "features": [
    {
      "id": 1,
      "type": "Feature",
      "geometry": {
        "type": "LineString",
        "coordinates": [[-73.603118,45.446466], [-73.543133, 45.582736], [ -73.535376, 45.596572]]
      },
      "properties": { .... }
    },
    { ... 2nd feature }
  ]
}
```

The result of the database query now is multiple rows of features:

```
row 1:  {
    "id": 1,
    "type": "Feature",
    "geometry": {
      "type": "LineString",
      "coordinates": [[-73.603118,45.446466], [-73.543133, 45.582736], [ -73.535376, 45.596572]]
    },
    "properties": { .... }
  }
row 2:  { ... 2nd feature }
```