Add a weight for bicycle=discouraged tag in OSRM profile
In general the OSM community do not encourage the use for this tag, but it can
be a good indication of dangerous bike routes. (We have a few street marked as such in Montreal)

This add a penalty to routing on those segments.

This was tested on De La Verandrye street, after setting the secondary default weight to 0.9
to have OSRM first route on it. After adding this code with the 0.9 weight, we see
that the road is avoided.