Rework pyTransition README to integrate actual example scripts
We add a python script (code-to-readme.py, created by Claude.AI) which will replace
marked script sections with actual content of scripts.

This is driven by a configuration files, readme-examples-config.yaml which define
name of script file and section of the script to extract.

We need to run the following to update the README file:
python3 code-to-readme.py --readme README.md --config readme-examples-config.yaml

Documentation about this script usage and the config file format is in the script header

We extracted 4 scripts from the previous README and put them in the examples directory.
Scripts errors were fixed