Add an TripTimeChoice enum to python library to validate the parameter departure_or_arrival_choice
Adding the enum insure that we don't fiddle with string and risk miswriting them. We now have a
clear DEPARTURE and ARRIVAL type. Since we use the same value as before for the enum, script
that will use the right strings will still work

Updated one of the example and added unittests

Issue: #1177