paths: update layers when TransitPathEdit component is mounted
fixes #1178

The upgrade to React 19 seemingly changed the handling of events in such
a way that the path selection event is now handled before the
`TransitPathEdit` component is mounted, while it was handled after in
the React 16 version. So we call the `updateLayers` method at mount to
display the selected path.