Wrap raise_for_status to check content response on 400 error
The Transition API define that the response content have the error details.
By default, the raise_for_status function check for the HTTP Reason field.
(This is an optional field where the server can add a phrase to details the
error.)
Instead of changing the way we send the error information, we wrap the check
and append the error message to the reason field in the response object.
We keep the same HttpError exception, but have a more verbose message.

Now we have, for example:
requests.exceptions.HTTPError: 400 Client Error: tripRoutingQueryAttributes validateParams:
timeSecondsSinceMidnight should be a positive integer for url: http://localhost:8080/api/v1/route?withGeojson=true