Fix 1_downloadOsmDataToFiles task.
The download Osm data task worked incorrectly in the step where the geojson is written to a file, resulting in an incorrectly formatted geojson.
This commit changes what is written to the file so that the json is properly formatted and has closing brackets in the right place.
We also put the stream writing in a promise and add a check that pauses the reading when the writing isn't done yet, so the two stream remain in sync when writing a large file.
Finally, we also replace the promisify(pipeline) with just a pipeline that comes from the node:stream/promises module, which does the same thing for less lines of code.