Stream enhance polygons task.
The 1b_prepareAndEnhanceOsmPolygonData task now works by streaming, reading the source files piece by piece and writing the result piece by piece to the destination file.
This does not change the result, but works better with big data, reducing memory usage and allowing for the treatment of a large geojson file without crashing.