dep: Upgrade react-datepicker to 7.6.0
fixes #1209

Also, use the default `css` file from the react-datepicker package
instead of overwriting the whole file and updating a few items. If we
need to overwrite some specific styles, it should be done only for the
changes to do, not copying the hwle file.

Update snapshots for the `Calendar` component.