Add option to getGeojsonsFromRawData()
Modifies the function getGeojsonsFromRawData() so that it accepts a new option parameter, continueOnMissingGeojson.
When generateNodesIfNotFound is false and continueOnMissingGeojson true, the function will just go to the next geojson and print a warning if a feature that is in the osm raw data is not in the geojson data.
Previously, the only option was to throw an error and interrupt the process.