config: move the `maxParallelCalculators` config to backend
This value should be set only on the server side and send on request to
the frontend for validations (in the batch calculation forms). It is
always available in the server config and it defaults to the number of
available CPUs.

Update the unit tests to test the default value. The `config2_test.js`
does not initialize this value anymore.