batch: use `maxParallelCalculations` when count is unset
fixes #1083

The previus `cpuCount` and `maxCpuCount` of the
`batchCalculationParameters` are replaced with a single optional
`parallelCalculations` parameters. If this parameter remains unset, the
batch routing job will use the `maxParallelCalculations` value.

This will allow currently running batch job to be continued on a server
with more CPUs available without having to recreate the job and set the
max value again.