path: use a waypoint/node property for errors instead of layers
The `transitNodesSelectedErrors` and `transitPathWaypointsErrors` layers
are removed. Instead the `transitNodesSelected` and the
`transitPathWaypoints` layers' features have an additional boolean
property that will be used to specify the strokeColor. This simplifies
those layer managements and ensure all waypoints and nodes actions work
identically whether the object is in error or not.