deck.gl: add a stroke layer to animated path
Add a `stroked` property to the `BaseLineLayerConfiguration` type.
Defaults to `false`.

Change the return type of the `getLayer` to an array, as the outline
will be an additional layer.

If the `stroked` is true, it adds an additional line layer that is 2
pixels wider than the main layer and of white color.