auth: type the `google` and `facebook` auth config
These options are optional boolean options added to the `auth` object of
the project.config's file.

Also hide the corresponding routes and auth methods conditional to the
enablement of those methods instead of being available all the time.

Display an error message if one of those auth method is enabled but the
required environment variables are not set.