Add new comparison tab
Starts phase 2 of the Habitation project by adding a new tab to transition that allows us to compare the same OD pair with different scenarios.
This reuses and modifies the routing tab, adding a second scenario selection for the same OD pair.
We can then see the two different paths on the same map, as well as a table that compares stats for the two trips with the different scenarios.

This adds a new layers to the map, which allows us to display a second trajectory with a differently colored outline, corresponding to the second selected scenario.
Text corresponding to the second scenario also has that different color to help the user link the two.
A few existing components, such as the route step buttons,  have also been modified to allow for a different text color than the default in the UI.