map: implement a custom controller to handle map events
The `TransitionMapController` will replace the default controller, so
that it is possible to handle events that are not exposed by Deck's
default API (for example the `pointermove` event, which will be useful
for the polygon selection tool).

The controller receives the active event manager, as well as map
callbacks to access some methods from the map to pick the objects under
the pointer or translate pixel coordinates to geometric coordinates.