map: set the `handled` state of the events and stop propagation
When a event handler has successfully run and handled the event, it should
return `true` and the caller needs to set the `handled` property of the
event to `true` to avoid other handlers running.

All event handlers now need to return a boolean value. This prevents the
map controller's event handler to run when the event has already been
handled by a layer.

The propagation is not stopped for dragStart/dragEnd as otherwise the
Deck.GL map does not properly handle the dragging state.