map: add a polygon draw tool to the main map
The draw tool is one of the edit tools. It draws a polygon as the user
clicks on the map, double-clicking will close the polygon and trigger
the `map.draw.polygon` event that receives an object with a `polygons`
field containing  polygon feature collection.

It appears in the main map only when the 'nodes' section is active, for
now, as it was this way before.