nodes: bring back the multiple node selection feature
When a polygon is drawn the nodes in that polygon are selected. The
functions that were previously in the `TransitionMainMap` are now moved
to the `TransitNodePanel`, closer to where they should be applied.

Context menu callbacks are updated accordingly.

A new layer has been added in the node section: the
`transitNodesSelectedPolygon`, which contain the polygon that was drawn
by the user using the DrawPolygon tool.