gtfs: generate paths even when it has no shape
fixes #275

This generates the path from the stop_times data and the stop
coordinates. It creates direct lines between those points and distances
are not calculated. This will allow to route using those paths.

Paths are considered identical if they have the same stop sequence.

Warnings are shown for each path of a line when it has no shape.