services: add back number of lines in service list
This shows the number of lines that are used by a service next to the
title. Also only display the lines if there actually are lines for the
service.

Display the `TransitServiceLinesDetails` only if it is opened. On some
instances with a lot of services, the service list was taking too much
time and memory to compute and was making the view unresponsive. So we
display this component only when it is required.