osrm: only return known and typed data in route results
Part of #1263

OSRM's API returns quite a lot of data with a route request, data that
Transition does not make any use of currently and that cannot be
generalized to our `RoutingResult` type. To avoid cluttering the result
object with unnecessary data, the `OSRMRoutingService` maps only the
known fields and drops other.

FIXME If we need any additional fields, we can type them in the
`Route` type in a engine independent way.

To give ballpark numbers, simply by doing this, a 8km walking trip route
result dropped from 48K to 18K.