batch: Do not save geometries if they are not requested
fixes #1263

Geometries can take a lot of space in the results data. So if the
user did not explicitly ask for them in the unimodal modes, they are
deleted from the response object.

To give ballpark numbers, for an 8km walking path, the size of the json
response has gone from 18K to 1K (without transit results).