update selected objects class and fix events problems in forms
This is a refactoring of the selected objects manager. Alongside this refactoring, the calls to emeit and events listeners have been updated to fix problems with events not having been called in the correct order since react 19 update.
ContextMenu in MainMapState was creating a new root for each render, which was causing problems with the context menu not being displayed. Now it is created once and reused.