Accessibility map comparison
Continues phase 2 of the Habitation project by adding a new tab to transition that allows us to compare accessibility maps centered on the same point but with different scenarios.
This reuses and modifies the accessibility map tab, adding a second scenario selection for the same origin.
The two accessibilty polygons are displayed on the same map, with different colors two represent the two scenarios, as well as their intersection.
In the case of more than one polygon being calculated, a dropdown menu allows for the selection of different max times.

The icon used for the new tab comes from https://www.svgrepo.com/svg/533376/diagram-venn, under a CC Attribution License.
We have also added a section to the Readme of our SVG icons' directory that lists the sources of our icons.