Add current calculation rate to log message
When running batch routing, we have a log which show the rate of calculation since
the beginning of the run. Since the rate can be non constant, we want to see
what is the current rate.
We calculate the rate between 2 log entries, by saving what were the values when we
previously displayed the log.

New log:
 calc/sec: 2.23 (current: 2.23)
 calc/sec: 3.14 (current: 4.91)
 calc/sec: 3.35 (current: 3.8)

Closes: #1262