Fix Routing Comparison so that form inputs affect both routing engines.
Some form parameters only affect the engine for the first scenario, not the second.
For example, if the starting time was changed, the first path would take the new time into account but not the second.
This commit adds a new updateBothROutingEngines method so that the new parameter value is applied to both engines.