osm: add task to import poi and residential data in database
fixes #1289

This tasks is the last task of the OSM import that was missing, to
import the generated and cleaned data saved in a geojson file to the
database.

This adds a task, `4_importPlacesToDb.task.ts` in `transition-backend`
that will automatically read the POIs and residential entrances file and
create/update 2 data sources, identified by the data source direction,
followed by ' Places' and ' Residences' respectively.

This task uses an import function that takes a point geojson file and
data source name and does the actual import, asking the user whether to
overwrite any previous data source of the same name. This import
function can later be re-used by other code paths to import places.