config: get the section configuration out of preferences
fixes #1286

Add a `sections` configuration option in the config of chaire-lib, to be
used by applications like Transition or Octavi that use the dashboard
functionality of chaire-lib.

Add a `project.config` file in transition-common with the transition
section default configuration.

Update the `webpack.config.js` file to make sure the default section
config is imported before reading the config file. That is the only
change required to account for the new section config option.

Update usages in the frontend to get the enabled sections from the
config instead of preferences.

** Breaking change **

Section configurations on the config file's `defaultPreferences` section
is now deprecated. They will still work for a while, but they need to be
updated.

Configuration files can overwrite section configuration options,
directly, not in the preferences, as follows:

```
module.exports = {
   ...myConfig,
   sections: {
      simulations: {
         enabled: true
      },
      routing: {
         enabled: false
      }
   }
}
```