Fix bug with passwordless login form when pressing 'enter' or 'space'. (#1302)
Fix: #725
Remove onKeyUp on the button because the default button already handles that.
On the email input, put onKeyUp with handleKeyUp.
So now, the user can press 'enter' or 'space' directly from the email input and go to the survey directly.
Remove e.witch that is deprecated. The new method is e.key.