Fix double submit when 'Enter' press inside the form. (#1304)
We remove onKeyUp inside a form element because this could send the form twice when clicking 'enter' with the submit button focus.
We move onKeyUp on the last input of the form so the user can submit the form when pressing 'enter' from there.
We did this correction for LoginForm, RegisterForm, and PwdLessLoginForm. There were the same errors for all of them.
Update the packages/chaire-lib-frontend/src/components/pages/__tests__/__snapshots__/LoginPage.test.tsx.snap snapshot with buttons with type 'submit'