Move json2capnp serialization code to a sub library
We move all the code that specificially handle writing and reading the capnp into a standalone
library. The aim of this is to eventually call this library directly instead of passing
through a daemon.

The main application keeps all the http handling and protocol decoding. We keep the various
router file, as they contains tests that are based on an end to end http communication.

All the capnp handling is moved to a "transition_capnp_data" library. This basically only
have the file writing and handling function. We did not duplicate the test at the moment since
the http based one still cover all the code.