dep: Upgrade socket.io to 4.8.1
Upgrade the `socket.io` and `socket.io-client` packages to version 4.8.1

Update the socketServerApp startup with the new types and make it
compile.

The main change resides in the socket.io-file server and client packages
that do not work with latest socket.io. The file upload needs to be
updated and done manually as per the documentation at https://socket.io/how-to/upload-a-file
and adapted to support our use-cases for import.

On the client side `FileUploaderHOC` is replaced by the
`FileUploaderHook` that is used by all importers, ie `FileImportForm`,
`ConfigureDemandFromCsvForm`, `GtfsImportForm` and
`AccessibilityMapBatchForm`.