gtfs import form: Convert to function component
This will allow this component to use a file upload hook instead of
re-implementing the file upload client logic in the class, when we drop
the socket.io-file-client in the 4.x migration.

The `GtfsValidator` object is not required anymore and its content is
rather used as different states in the form.