Disable useShell for starting trRouting
The memcached support have changed the start of trRouting to be in a Shell.
it could not connect to memcached in a docker container without it.
Alas, this make it impossible to kill trRouting to be stop/restart with the current method.

Let's move back to useShell=false. This will have trRouting fail to connect to
memcached, but that's ok. It still work, but warn about the failed memcached connection.