pathCollection: add a `toGeojsonSimplified` function
This creates a geojson feature collection that contains only the
properties that are required for the display of paths. Since this data
can be copied to the graphic card, to have smaller data will make the
operation less memory intensive, especially that paths have a lot of
additional data properties that are not required for display purposes.