Routing and accessibility map tabs: Remove error messages when form values are changed.
The accessibility maps and routing tabs do not remove routing errors, even after switching to new valid values.
This commit empties the routing errors when changing a form value so that the error won't be displayed anymore.

This also removes the error message about invalid scenarios when switching back to a tab with a selected scenario after deleting said scenario.
It will now select the empty choice while giving a "The scenario is required." error.

Fix: #1331