Remove the id attribute of userEmailData
Since this attribute is only used to display an error message that is prone to misinterpretation when the email is null or undefined, we have removed it.
The email attribute's type was changed to just be string as a result, and the type verification has been moved to a new function that is called before sendEmail().
Fix: #836