user: Add possibility to set users as admin
fixes #654

Add a route in the backend such that the is_admin flag can only be set
by an administrator. Extract the user update code so that the other user
update route can use the same code. It sanitizes this other route to
make sure the is_admin flag cannot be set by a non-admin user.

Add a component to edit the admin status of a user in the users table.
A user cannot set his own administrator flag, to not lock himself out of
the application.