preferences: Make sure `infoPanelPosition` cannot be blank
fixes #1362

Add a `noBlank` prop to the `infoPanelPosition` preferences selection.
Also, the value will be shown as left only if the value of the
preference is `left`, otherwise, the right will be shown.

In the `TransitionDashboard` component, since we have only 2 possible
positions, left or right, we put the panel on the left if and only if
the value of the preference is left. Otherwise, we put it on the right.
No instead of checking equality with `right`, we check inequality with
`left`. That can be revisited when we support more placement options.