privateRoute: do not return different object in useSelector
There is this warning in the console: `Selector unknown returned a
different result when called with the same parameters. This can lead to
unnecessary rerenders.` Error message points to this page for
explanations https://redux.js.org/usage/deriving-data-selectors#writing-memoized-selectors-with-reselect

To avoid unnecessary re-renders, the function in `useSelector` should
avoid doing any processing on the values and simply return the necessary
data or memoize the returned value.

In Transition the `PrivateRoute` component is the only one causing this
warning and it is fixed by calling the `useSelector` function once for
each desired value.