tests: destroy connection after db tests
By adding `await knex.destroy()` in the `afterAll` block of each
database tests, we make sure that the connection is properly destroyed,
synchronously, and the tests can exit without open handles.

Some tests already called the `destroy` db query from some tables, but
this function was effectively destroying the connection multiple times
(as all queries share the same connection) and it was not done
synchronously as no callback was passed to the function, so the test was
still finishing with open handles.