auth: pass a router to setup auth routes instead of app
fixes #1392

Setting the routes on a router gives more flexibility to the caller to
add middleware and prefix the paths to the auth routes.

The app is a `Router`, but the reverse is not true. So we do not need to
change the way those routes are set in the Transition app, as the
express application in Transition is quite simple.