jobs: set status to `failed` if completed is `false`
Currently, if the job's function returns correctly, the job is always
set to completed, even if the `completed` value of the results is
`false` because there were parsing errors in the file.

We look at the `completed` field before setting the status to
`completed` or `failed`.

Add unit tests for the executable job execution in the workerpool. We
need to export the `wrapTaskExecution` to be able to unit test it.