captcha: refactor captcha component to support many types
Add a project config option to specify the type of captcha. For now, the
captcha has `simple` type that is an image from which the user should
identify the characters.

Move the `CaptchaComponent` to a captcha directory as it is not
necessarily used only for local login.

Rename `CaptchaComponent` to `simpleCaptcha` and add a
`CaptchaComponent` that has the responsibility to load the requested
captcha when we support more.