captcha: Add a captcha component with @cap.js
Add routes in the backend to support the `@cap.js/server` challenge and

Add the `capjs` type of captcha in the project configuration type.
redeem. Those routes are conditional on the requested captcha type in
the backend and are automatically loaded with then authentication
routes. The `captcha.routes` file also provides a middleware that can
check the validity of a captcha token.

Add a captcha component using the `@pitininja/cap-react-widget` package.