captcha: allow to keep a token after validation
By default, once a cap.js token is validated, it is deleted and cannot
be reused by any other query. This is not desirable, for example, in
case of login, where legitimate errors can occur and the user may need
to re-submit a modified form. He should not have to re-do the captcha
every time.

This adds an option to the captcha middleware, a boolean `keepToken`
value, which can be set per route, such that a query that should succeed
everytime (like help request) can delete the token, while others, like
login, can keep the token between queries.

Note that cap.js does a garbage collection of the expired token in any
case, this option only applies to the current query.

Also reset the capjs captcha when the `reloadKey` is updated in the
React component.