db: filter lines, agencies, services and path collection by scenario
Add a `scenarioId` optional parameter to the `collection` functions of
the lines, agencies and services db queries. When the scenario is set,
it returns only the objects for which there is service on this scenario.

It also filters on `only_agencies, `only_lines`, `except_agencies` and
`except_lines` if these values are set.

Support inclusions and exclusions in the path geojson collection query
by scenario ID.