db: filter lines, agencies, services and path collection by scenario
Add `collectionForScenario` functions in the db queries for agencies,
lines, services and paths. It returns only the objects for which there
is service on this scenario.

It also filters on `only_agencies, `only_lines`, `except_agencies` and
`except_lines` if these values are set.

In paths db queries, filter by scenario was already supported, but the
call with scenarioId has been deprecated in favor of the
`geojsonCollectionForScenario`, which now supports inclusions and
exclusions.

Add scenario data for database tests and exhaustively test these new
functions.