scenarioEdit: use the `transitPaths.getForScenario` socket route
fixes #667

This socket route, defined, but not used yet, will now fetch the paths
for a specific scenario, including exclusions and inclusions, instead of
only the paths for the selected services.

If the scenario is new though, the paths for the services are still
fetched, as the route with the scenario ID will not work since the
scenario is not in the DB yet.