nodes: add a query to get transferable nodes between paths
This query takes 2 arrays of paths: one for the unboarding nodes and
another for the boarding nodes. It will find all the pairs of nodes that
can transfer from one path to the other, returning the transfer travel
time and distance.

If many nodes are reachable from a path/node pair, it will return the
one with the smallest time to reach.