routing validation: add frontend component to configure task
Update the batch calculation panel to support both the routing
calculation and validation calculation from csv file. Both will be in
separate tabs of the panel

Add types for the validation demand CSV file. It is the same as the
OdDemand, except it adds a trip date and agencies and lines field
prefix.