Import census canada data
Adds a task that reads two files provided by census canada, adding their info to the databse.
The first is a gml that contains the geographic boundaries of every dissemination block in canada, the smallest division used in the census.
The second is a csv file that has the population of these blocks, as well as proximity indices that indicate the access to categories of services from each of these blocks.
The task reads and adds the info for all 489546 blocks, taking about 250 seconds on my computer.
Fix: #1396