Move the non-socket.io setup bits directly into server.ts
The routing and workerpool had not link to the socket.io setup, so we move them into the server.ts file.

Created a new function setupRouting that deal with OSRM, cache creation and trRouting. Reworked the await
so that we can start OSRM and recreate the cache in parallel for a small startup speed up.

Moved the code that was in server.ts into a setupAll() function, so that we can use await there. We
call the setupRouting and await at the end, before enqueuing the batch job. That way, we can start
the webapp and socket.io before the routing is fully setup. Since we wait after the routing await to enqueue
the batch job, this will not create a issue and solve #898. To insure a user cannot start a new batch job
before the end of the routing setup, we only start the worker pool after it's done.

Also moved the http server creation into a function to isolate from the rest.