Fix yarn format command in macos
In macos, when there is no quotes around the prettier path,
it ignores most files and will not fix formatting in several subdirectories.
Using quotes seems to fix the issue.

In transition, there was PWD before the path with single quotes.
This was removed and replaced with double quotes and should work fine now.