Pass an ExecutableJob object to batchRouting and batchAccessibility function runners
Instead of having several parameters to manage for each functions, we pass a single
job object and we fetch the data from there. This will allow us to encapsulate other
aspect, like file access.

We add a few file management function to ExecutableJob.

We change the unit test to load an actual ExecutableJob instead of just mock data via
a mock of the db queries.