dep: replace `osmtogeojson` with `osm2geojson-lite`
fixes #1364

This is a drop-in replacement of the original package, with less
dependencies and more supported.

While the geojson generated is a bit different, the differences mostly
reside in how it handles relations. `osmtogeojson` would create the
features for each feature of the relation, creating duplicates in the
content, while `osmtogeojson-lite` will create a node where the relation
happens and a single feature for the relation itself.

For our use cases, that difference in geojson output does not pose
problems. Our import scripts and usages give the same results as
previously.