Move all tasks to the -backend packages
Since tasks run on the CLI, they can only be ran in the backend, so placing them all in the backend packages makes sense.

At the initial split, some of the tasks were kept in the -common packages, probably due to complex dependencies.

We extract some of the dependencies and move the code that was only used by the tasks also in the backend.