utils: Add a `BaseObject` abstract class for base objects
This class is part of a refactorigin to replace the `GenericObject` and
`ObjectWithHistory` with common object functionalities, that limit
possible errors and reduces the number of un-necessary fields in some
objects.

This base class handles the attributes handling, type safety and object
validation. It does not require any specific type of attributes, just
any `Record<string, unknown>`

This greatly limits the functionalities that were available in the
GenericObject class, the `_attributes` is now a private field and the
`attributes` getter is readonly. The attributes can only be set through
the `set` method which is correctly typed to make sure `path` and
`value` have the expected values. A `mergeAttributes` method allows to
set many attributes at the same time. This class also has a `validate`
method, which calls the abstract `_validate` method that child classes
should implement.