Let the `TransitRouting` class extend the `BaseObject`
The `TransitRoutingAttributes` are not to be saved to the database, so
it does not require all the default fields form the `GenericObject`
class.

Also, it does not support yet undoing/redoing, so the
`ObjectWithHistory` extension is un-necessary as well. It is a good
first candidate to extend the new BaseObject class.